+++ coda-src/venus/venus.private.h
@@ -341,7 +341,7 @@ extern const char *kernDevice;
 extern const char *realmtab;
 extern const char *CacheDir;
 extern const char *CachePrefix;
-extern unsigned int CacheBlocks;
+extern uint64_t CacheBlocks;
 extern const char *SpoolDir;
 extern uid_t PrimaryUser;
 extern const char *VenusPidFile;
